dante correcion semana 4

+++ src/main/java/pe/edu/upc/trabajofinalbackend/repositories/ICommentForumRepository.java
@@ -7,6 +7,6 @@ import pe.edu.upc.trabajofinalbackend.entities.CommentForum;
 import java.util.List;
 
 @Repository
-public interface ICommentForumRepository extends JpaRepository<CommentForum, Long> {
+public interface ICommentForumRepository extends JpaRepository<CommentForum, Integer> {
 
 }
